Move exports transformer into exports folder
