+++ src/apps/companies/transformers/index.js
@@ -1,10 +1,8 @@
-const transformCompanyToExportDetailsView = require('./company-to-export-details-view')
 const transformCompanyToListItem = require('./company-to-list-item')
 const transformCompanyToSubsidiaryListItem = require('./company-to-subsidiary-list-item')
 const transformCoreTeamToCollection = require('./one-list-core-team-to-collection')
 
 module.exports = {
-  transformCompanyToExportDetailsView,
   transformCompanyToListItem,
   transformCompanyToSubsidiaryListItem,
   transformCoreTeamToCollection,
